refactor(u): Move array serialization logic to u
